feat: Add new SortingTable component
