+++ src/SortingTable/index.ts
@@ -1,0 +1,1 @@
+export { default as SortingTable } from "./SortingTable";
